+++ medusa/clients/torrent/transmission.py
@@ -12,6 +12,7 @@ from base64 import b64encode
 
 from medusa import app
 from medusa.clients.torrent.generic import GenericClient
+from medusa.helper.exceptions import DownloadClientConnectionException
 from medusa.logger.adapters.style import BraceAdapter
 from medusa.schedulers.download_handler import ClientStatus
 
@@ -308,8 +309,7 @@ class TransmissionAPI(GenericClient):
         post_data = json.dumps({'arguments': return_params, 'method': 'torrent-get'})
 
         if not self._request(method='post', data=post_data) or not self.check_response():
-            log.warning('Error while fetching torrent {hash} status.', {'hash': info_hash})
-            return
+            raise DownloadClientConnectionException(f'Error while fetching torrent {info_hash} status.')
 
         torrent = self.response.json()['arguments']['torrents']
         if not torrent: