+++ medusa/server/core.py
@@ -29,6 +29,7 @@ from medusa.server.api.v2.internal import InternalHandler
 from medusa.server.api.v2.log import LogHandler
 from medusa.server.api.v2.postprocess import PostProcessHandler
 from medusa.server.api.v2.providers import ProvidersHandler
+from medusa.server.api.v2.recommended import RecommendedHandler
 from medusa.server.api.v2.schedule import ScheduleHandler
 from medusa.server.api.v2.search import SearchHandler
 from medusa.server.api.v2.series import SeriesHandler
@@ -140,6 +141,9 @@ def get_apiv2_handlers(base):
 
         # /api/v2/authenticate
         AuthHandler.create_app_handler(base),
+
+        # /api/v2/recommeded
+        RecommendedHandler.create_app_handler(base),
 
         # Always keep this last!
         NotFoundHandler.create_app_handler(base)