+++ themes-default/slim/src/router/routes.js
@@ -255,6 +255,7 @@ const addShowRoutes = [
             topMenu: 'home',
             converted: true
         },
+        props: route => ({ ...route.params }),
         component: () => import('../components/new-show.vue')
     },
     {
@@ -304,7 +305,7 @@ const addRecommendedRoute = {
         topMenu: 'home',
         converted: true
     },
-    component: () => import('../components/add-recommended.vue')
+    component: () => import('../components/recommended.vue')
 };
 
 /** @type {import('.').Route} */