+++ themes-default/slim/src/store/modules/config/indexers.js
@@ -6,7 +6,8 @@ const state = {
         statusMap: {},
         traktIndexers: {},
         validLanguages: [],
-        langabbvToId: {}
+        langabbvToId: {},
+        recommendedLists: {}
     },
     indexers: {
         tvdb: {